chore: move generic params helper into react helpers
